Refactoring select all, move code to hooks, removed duplicated state from store
