+++ src/features/calendar/components/CalendarWeekView/index.tsx
@@ -1,3 +1,4 @@
+import { useEffect, useRef, useState } from 'react';
 import { Box } from '@mui/system';
 import dayjs from 'dayjs';
 import { FormattedTime } from 'react-intl';
@@ -11,25 +12,24 @@ import {
   Typography,
   useTheme,
 } from '@mui/material';
-import { useEffect, useMemo, useRef, useState } from 'react';
 
+import { isSameDate } from 'utils/dateUtils';
+import messageIds from 'features/calendar/l10n/messageIds';
+import useCreateEvent from 'features/events/hooks/useCreateEvent';
+import useWeekCalendarEvents from 'features/calendar/hooks/useWeekCalendarEvents';
+import { useFocusDate } from 'utils/hooks/useFocusDate';
+import { useWeekDates } from 'features/calendar/hooks/useWeekDates';
+import { Msg } from 'core/i18n';
+import { useNumericRouteParams } from 'core/hooks';
+import range from 'utils/range';
 import DayHeader from './DayHeader';
 import EventCluster from '../EventCluster';
 import EventDayLane from './EventDayLane';
 import EventGhost from './EventGhost';
 import EventShiftModal from '../EventShiftModal';
 import HeaderWeekNumber from './HeaderWeekNumber';
-import { isSameDate } from 'utils/dateUtils';
-import messageIds from 'features/calendar/l10n/messageIds';
-import { Msg } from 'core/i18n';
-import range from 'utils/range';
 import { scrollToEarliestEvent } from './utils';
 import { getDstChangeAtDate } from '../utils';
-import useCreateEvent from 'features/events/hooks/useCreateEvent';
-import { useAppDispatch, useNumericRouteParams } from 'core/hooks';
-import useWeekCalendarEvents from 'features/calendar/hooks/useWeekCalendarEvents';
-import { setWeekViewDates } from 'features/calendar/store';
-import { useFocusDate } from 'utils/hooks/useFocusDate';
 
 dayjs.extend(isoWeek);
 
@@ -41,7 +41,6 @@ export interface CalendarWeekViewProps {
 }
 const CalendarWeekView = ({ onClickDay }: CalendarWeekViewProps) => {
   const theme = useTheme();
-  const dispatch = useAppDispatch();
   const [creating, setCreating] = useState(false);
   const [shiftModalOpen, setShiftModalOpen] = useState(false);
   const [pendingEvent, setPendingEvent] = useState<[Date, Date] | null>(null);
@@ -51,33 +50,14 @@ const CalendarWeekView = ({ onClickDay }: CalendarWeekViewProps) => {
   const { orgId, campId } = useNumericRouteParams();
   const createEvent = useCreateEvent(orgId);
   const { focusDate } = useFocusDate();
-  const focusWeekStartDay = useMemo(() => {
-    return dayjs(focusDate).isoWeekday() == 7
-      ? dayjs(focusDate).add(-1, 'day')
-      : dayjs(focusDate);
-  }, [focusDate]);
+  const { weekDates } = useWeekDates();
 
-  const weekViewDates = useMemo(() => {
-    return range(7).map((weekday) =>
-      focusWeekStartDay.day(weekday + 1).toDate()
-    );
-  }, [focusWeekStartDay]);
-
-  useEffect(() => {
-    dispatch(setWeekViewDates(weekViewDates.map((d) => d.toISOString())));
-  }, [weekViewDates]);
-
-  const dstChange = useMemo(
-    () =>
-      weekViewDates
-        .map((d) => dayjs(d))
-        .find((date) => getDstChangeAtDate(date)),
-    [weekViewDates]
-  );
+  const dstChange = weekDates
+    .map((d) => dayjs(d))
+    .find((date) => getDstChangeAtDate(date));
 
   const eventsByDate = useWeekCalendarEvents({
     campaignId: campId,
-    dates: weekViewDates,
     orgId,
   });
 
@@ -105,8 +85,8 @@ const CalendarWeekView = ({ onClickDay }: CalendarWeekViewProps) => {
         position="relative"
       >
         {/* Empty */}
-        <HeaderWeekNumber weekNr={dayjs(weekViewDates[0]).isoWeek()} />
-        {weekViewDates.map((weekdayDate: Date, weekday: number) => {
+        <HeaderWeekNumber weekNr={dayjs(weekDates[0]).isoWeek()} />
+        {weekDates.map((weekdayDate: Date, weekday: number) => {
           return (
             <Box key={`weekday-${weekday}`} position="relative">
               <DayHeader
@@ -169,7 +149,7 @@ const CalendarWeekView = ({ onClickDay }: CalendarWeekViewProps) => {
           })}
         </Box>
         {/* Day columns */}
-        {weekViewDates.map((date: Date, index: number) => {
+        {weekDates.map((date: Date, index: number) => {
           const pendingTop = pendingEvent
             ? (pendingEvent[0].getUTCHours() * 60 +
                 pendingEvent[0].getMinutes()) /
