+++ src/features/calendar/store.ts
@@ -5,23 +5,13 @@ import { TimeScale } from './types';
 export interface CalendarStoreSlice {
   focusDate: string;
   maxMonthEventsPerDay: number;
-  weekViewDates: string[];
   timeScale: string;
-  monthViewSpan: {
-    endDate: string | undefined;
-    startDate: string | undefined;
-  };
 }
 
 const initialState: CalendarStoreSlice = {
   focusDate: new Date().toISOString(),
   maxMonthEventsPerDay: 0,
-  monthViewSpan: {
-    endDate: undefined,
-    startDate: undefined,
-  },
   timeScale: '',
-  weekViewDates: [],
 };
 
 const calendarSlice = createSlice({
@@ -34,30 +24,12 @@ const calendarSlice = createSlice({
     setMaxMonthEventsPerDay(state, action: PayloadAction<number>) {
       state.maxMonthEventsPerDay = action.payload;
     },
-    setMonthViewSpan(
-      state,
-      action: PayloadAction<{ endDate: string; startDate: string }>
-    ) {
-      const { startDate, endDate } = action.payload;
-      state.monthViewSpan = {
-        endDate: endDate,
-        startDate: startDate,
-      };
-    },
     setTimeScale(state, action: PayloadAction<TimeScale>) {
       state.timeScale = action.payload;
-    },
-    setWeekViewDates(state, action: PayloadAction<string[]>) {
-      state.weekViewDates = action.payload;
     },
   },
 });
 
 export default calendarSlice;
-export const {
-  setFocusDate,
-  setMaxMonthEventsPerDay,
-  setMonthViewSpan,
-  setTimeScale,
-  setWeekViewDates,
-} = calendarSlice.actions;
+export const { setFocusDate, setMaxMonthEventsPerDay, setTimeScale } =
+  calendarSlice.actions;
