+++ src/features/events/components/SelectionBar/index.tsx
@@ -9,93 +9,24 @@ import {
   Paper,
   Typography,
 } from '@mui/material';
-import dayjs from 'dayjs';
 
+import { eventsSelected, resetSelection } from 'features/events/store';
+import useParticipantPool from 'features/events/hooks/useParticipantPool';
+import { RootState } from 'core/store';
+import { useAppDispatch, useAppSelector } from 'core/hooks';
+import { Msg } from 'core/i18n';
+import useVisibleEventIds from 'features/calendar/hooks/useVisibleEventIds';
 import EventParticipantsModal from '../EventParticipantsModal';
 import messageIds from '../../../calendar/l10n/messageIds';
 import MoveCopyButtons from './MoveCopyButtons';
-import { Msg } from 'core/i18n';
-import { eventsSelected, resetSelection } from 'features/events/store';
-import { RootState } from 'core/store';
 import SelectionBarEllipsis from '../SelectionBarEllipsis';
-import useParticipantPool from 'features/events/hooks/useParticipantPool';
-import {
-  useAppDispatch,
-  useAppSelector,
-  useNumericRouteParams,
-} from 'core/hooks';
-import useMonthCalendarEvents from 'features/calendar/hooks/useMonthCalendarEvents';
-import useWeekCalendarEvents from 'features/calendar/hooks/useWeekCalendarEvents';
-import { useFocusDate } from 'utils/hooks/useFocusDate';
-import useDayCalendarEvents from 'features/calendar/hooks/useDayCalendarEvents';
 
 const SelectionBar = () => {
   const dispatch = useAppDispatch();
   const [participantsDialogOpen, setParticipantsDialogOpen] = useState(false);
   const { affectedParticipantIds } = useParticipantPool();
-  const calendarStore = useAppSelector((state: RootState) => state.calendar);
-  const { focusDate } = useFocusDate();
 
-  const { orgId, campId } = useNumericRouteParams();
-  const monthViewSpan = useMemo(
-    () => ({
-      endDate:
-        calendarStore.monthViewSpan.endDate !== undefined
-          ? dayjs(calendarStore.monthViewSpan.endDate).toDate()
-          : undefined,
-      startDate:
-        calendarStore.monthViewSpan.startDate !== undefined
-          ? dayjs(calendarStore.monthViewSpan.startDate).toDate()
-          : undefined,
-    }),
-    [calendarStore.monthViewSpan]
-  );
-  const weekViewDates = useMemo(
-    () => calendarStore.weekViewDates.map((d) => dayjs(d).toDate()),
-    [calendarStore.weekViewDates]
-  );
-
-  const monthCalendarEvents = useMonthCalendarEvents({
-    campaignId: campId,
-    endDate: monthViewSpan.endDate,
-    maxPerDay: calendarStore.maxMonthEventsPerDay,
-    orgId,
-    startDate: monthViewSpan.startDate,
-  });
-
-  const weekCalendarEvents = useWeekCalendarEvents({
-    campaignId: campId,
-    dates: weekViewDates,
-    orgId,
-  });
-
-  const { activities: dayCalendarEvents } = useDayCalendarEvents(focusDate);
-
-  function getIdsToSelect() {
-    const eventsToSelect = new Set<number>();
-    if (calendarStore.timeScale === 'month') {
-      monthCalendarEvents
-        .flatMap((event) => event.clusters)
-        .flatMap((cluster) => cluster.events)
-        .map((event) => event.id)
-        .forEach((id) => eventsToSelect.add(id));
-    } else if (calendarStore.timeScale === 'week') {
-      weekCalendarEvents
-        .flatMap((events) => events.lanes)
-        .flatMap((lanes) => lanes)
-        .flatMap((lane) => lane.events)
-        .map((event) => event.id)
-        .forEach((id) => eventsToSelect.add(id));
-    } else if (calendarStore.timeScale === 'day') {
-      dayCalendarEvents
-        .flatMap((event) => event[1])
-        .flatMap((e) => e.events)
-        .map((e) => e.data.id)
-        .forEach((id) => eventsToSelect.add(id));
-    }
-    return Array.from(eventsToSelect);
-  }
-
+  const visibleEventIds = useVisibleEventIds();
   const selectedEventIds = useAppSelector(
     (state: RootState) => state.events.selectedEventIds
   );
@@ -105,12 +36,10 @@ const SelectionBar = () => {
   };
 
   const handleSelectAll = () => {
-    const eventIds = getIdsToSelect();
-    dispatch(eventsSelected(eventIds));
+    dispatch(eventsSelected(visibleEventIds));
   };
 
   const canSelectAll = useMemo(() => {
-    const visibleEventIds = getIdsToSelect();
     if (selectedEventIds.length !== visibleEventIds.length) {
       return true;
     }
@@ -118,12 +47,7 @@ const SelectionBar = () => {
       selectedEventIds.includes(v)
     );
     return canSelect;
-  }, [
-    selectedEventIds,
-    monthCalendarEvents,
-    weekCalendarEvents,
-    dayCalendarEvents,
-  ]);
+  }, [selectedEventIds]);
 
   return (
     <Box
